Move webstudio API tests under tests.api.webstudio
These tests never touched the UI — they were sitting under tests.ui.webstudio
purely by accident. Move them to a proper api-package hierarchy:

  tests.api.webstudio.client       — new client-projects validation tests
    AbstractStudioCentralProjectsApi
    AbstractZippedProjectsApi
    TestStudioCentralGroupRatingClaim
    TestStudioCentralGroupPolicyBundle
    TestZippedProjects
  tests.api.webstudio.integration  — older endpoint integration tests
    AbstractProjectApiTest
    TestProjectModulesRestApi
    TestProjectRunRestApi

Update the three suite XMLs (studio_central_projects_regression,
studio_zip_projects_regression, studio_api_integration) to point at the new
fully-qualified class names. The legacy UI test TestLocalZippedProjects
stays under tests.ui.webstudio.client.
