+++ src/test/resources/testng_suites/studio_zip_projects_regression.xml
@@ -6,6 +6,6 @@
         <listener class-name="configuration.listeners.ZippedProjectsReportPortalListener"/>
     </listeners>
 
-    <test name="TestZippedProjects"><classes><class name="tests.ui.webstudio.client.TestZippedProjects"/></classes></test>
+    <test name="TestZippedProjects"><classes><class name="tests.api.webstudio.client.TestZippedProjects"/></classes></test>
 
 </suite>
